requirements.txt
cachetools
certifi
charset-normalizer
click
Flask
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
httplib2
idna
itsdangerous
Jinja2
MarkupSafe
oauthlib
protobuf
pyasn1
pyasn1-modules
pyparsing
requests
requests-oauthlib
rsa
six
uritemplate
urllib3
Werkzeug
